% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dre_genes-data.R
\docType{data}
\name{Dre_genes}
\alias{Dre_genes}
\title{Dre_genes}
\format{
An object of class \code{data.frame} with 65905 rows and 4 columns.
}
\source{
\href{https://www.ensembl.org/index.html}{ensembl biomaRt}
}
\usage{
Dre_genes
}
\description{
A gene ID table for \emph{D. rerio}.
}
\details{
This table was built directly from the ensembl biomaRt.
}
\section{Format}{

A dataframe with the folowing columns :
\itemize{
\item \code{ens_id}: Ensembl ID (\emph{e.g} \code{ENSDARG00000069301})
\item \code{transcript_id}: Transcript ID (\emph{e.g} \code{ENSDART00000100744})
\item \code{public_name}: Gene name (\emph{e.g} \code{tmem177})
\item \code{transcript_length}: Transcript length from 5'UTR start to 3'UTR end (\emph{e.g} \code{2099})
}
}

\keyword{datasets}
